+++ bindings/vue/src/components/index.js
@@ -1,4 +1,5 @@
 export { default as VOnsButton } from './VOnsButton.vue';
 export { default as VOnsIcon } from './VOnsIcon.vue';
+export { default as VOnsInput } from './VOnsInput.vue';
 export { default as VOnsRange } from './VOnsRange.vue';
 export { default as VOnsSwitch } from './VOnsSwitch.vue';